Issue/29 (#30)
Functionality in Dotify so that a block that appears at the top of a page's content won't be displayed. We need this functionality for the Pipeline to become feasible not to display the contents of <pagenum> elements of a DTBook when they happen to be the first content on a page.